+++ keystore/store/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [keystore@1.0.0-canary.16](https://github.com/algorandfoundation/wallet-provider-extensions/compare/keystore@1.0.0-canary.15...keystore@1.0.0-canary.16) (2026-05-13)
+
+### Features
+
+- web crypto fallback for keys and generic secret import ([f7481b1](https://github.com/algorandfoundation/wallet-provider-extensions/commit/f7481b1dc95014c1903a4351d5fffc8c9995e31a))
+
 # [keystore@1.0.0-canary.15](https://github.com/algorandfoundation/wallet-provider-extensions/compare/keystore@1.0.0-canary.14...keystore@1.0.0-canary.15) (2026-04-16)
 
 ### Bug Fixes